refactor: Edit marketing, delete service
