+++ feature/mypage/src/main/java/com/teamoffroad/feature/mypage/data/repository/GainedEmblemsRepositoryImpl.kt
@@ -8,11 +8,11 @@ import com.teamoffroad.feature.mypage.domain.repository.GainedEmblemsRepository
 import javax.inject.Inject
 
 class GainedEmblemsRepositoryImpl @Inject constructor(
-    private val userService: EmblemService
+    private val emblemService: EmblemService
 ) : GainedEmblemsRepository {
 
     override suspend fun getGainedEmblems(): Result<List<GainedEmblems>?> {
-        val gainedEmblemResult = runCatching { userService.getGainedEmblems().data }
+        val gainedEmblemResult = runCatching { emblemService.getGainedEmblems().data }
         gainedEmblemResult.onSuccess { gainedEmblemsResponseDto ->
             val a = gainedEmblemsResponseDto?.gainedEmblems?.map {
                 it.toEmblemsList()
@@ -26,28 +26,6 @@ class GainedEmblemsRepositoryImpl @Inject constructor(
 
         gainedEmblemResult.onFailure {
             return Result.failure(it)
-        }
-        return Result.failure(UnReachableException("unreachable code"))
-    }
-
-    override suspend fun patchMarketingInfo(): Result<Unit> {
-        val patchMarketingInfoResult = runCatching { userService.patchMarketingInfo(true).data }
-        patchMarketingInfoResult.onSuccess {
-            Log.d("asdasd", it?.marketing?.data.toString())
-        }
-        patchMarketingInfoResult.onFailure {
-            Log.d("asdsad", it.message.toString())
-        }
-        return Result.failure(UnReachableException("unreachable code"))
-    }
-
-    override suspend fun deleteUserInfo(deleteCode: String): Result<Unit> {
-        val deleteUserInfo = runCatching { userService.deleteUserInfo(deleteCode) }
-        deleteUserInfo.onSuccess {
-            Log.d("deletesuccess", it.data.toString())
-        }
-        deleteUserInfo.onFailure {
-            Log.d("deletefail", it.message.toString())
         }
         return Result.failure(UnReachableException("unreachable code"))
     }
