+++ feature/mypage/src/main/java/com/teamoffroad/feature/mypage/domain/repository/GainedEmblemsRepository.kt
@@ -4,6 +4,5 @@ import com.teamoffroad.feature.mypage.domain.model.GainedEmblems
 
 interface GainedEmblemsRepository {
     suspend fun getGainedEmblems(): Result<List<GainedEmblems>?>
-    suspend fun patchMarketingInfo(): Result<Unit>
-    suspend fun deleteUserInfo(deleteCode: String): Result<Unit>
+
 }
