+++ feature/mypage/src/main/java/com/teamoffroad/feature/mypage/domain/repository/MyPageUserRepository.kt
@@ -4,4 +4,6 @@ import com.teamoffroad.feature.mypage.domain.model.MyPageUser
 
 interface MyPageUserRepository {
     suspend fun fetchMyPage(): MyPageUser
+    suspend fun patchMarketingInfo(): Result<Unit>
+    suspend fun deleteUserInfo(deleteCode: String): Result<Unit>
 }
