+++ feature/mypage/src/main/java/com/teamoffroad/feature/mypage/domain/usecase/DeleteUserInfoUseCase.kt
@@ -1,9 +1,9 @@
 package com.teamoffroad.feature.mypage.domain.usecase
 
-import com.teamoffroad.feature.mypage.domain.repository.UserRepository
+import com.teamoffroad.feature.mypage.domain.repository.MyPageUserRepository
 
 class DeleteUserInfoUseCase(
-    private val userRepository: UserRepository
+    private val userRepository: MyPageUserRepository
 ) {
     suspend operator fun invoke(deleteCode: String): Result<Unit> {
         return userRepository.deleteUserInfo(deleteCode)
