+++ feature/mypage/src/main/java/com/teamoffroad/feature/mypage/domain/usecase/UserMarketingInfoUseCase.kt
@@ -1,9 +1,9 @@
 package com.teamoffroad.feature.mypage.domain.usecase
 
-import com.teamoffroad.feature.mypage.domain.repository.UserRepository
+import com.teamoffroad.feature.mypage.domain.repository.MyPageUserRepository
 
 class UserMarketingInfoUseCase(
-    private val userRepository: UserRepository
+    private val userRepository: MyPageUserRepository
 ) {
     suspend operator fun invoke(): Result<Unit> {
         return userRepository.patchMarketingInfo()
